attribution param added to device info class
